+++ CHANGELOG.md
@@ -27,6 +27,7 @@
     * Fixed `header.admin.children_display_order` in Flex Pages to work just like with regular pages
     * Fixed `Utils::isFunctionDisabled()` method if there are spaces in `disable_functions` [#3023](https://github.com/getgrav/grav/issues/3023)
     * Fixed potential fatal error when creating flex index using cache [#3062](https://github.com/getgrav/grav/issues/3062)
+    * Fixed fatal error in `CompiledFile` if the cached version is broken
 
 # v1.7.0-rc.17
 ## 10/07/2020